Fix minor errors & add RunDiffToMaster
- Add new util task, RunDiffToMaster, to automatically run all modules that
  changed compared to master
- Fix minor issue with French Census table version number having been incremented
  sometime back when it didn't actually change
- Remove servicios_areas from MX MetaWrapper, as it doesn't actually have data
