+++ tests/test_catalog.py
@@ -1,19 +1,26 @@
 from tests.util import runtask, setup, teardown
-from tests.test_metadata import collect_meta_wrappers
+
+from tasks.util import TableTask
+
+# Monkeypatch TableTask
+TableTask._test = True
 
 from tasks.carto import OBSMetaToLocal
-from tasks.meta import current_session
-from tasks.util import MetaWrapper, TagsTask, ColumnsTask, TableTask
+from tasks.util import TagsTask, ColumnsTask, collect_meta_wrappers
 from tasks.sphinx import Catalog
 
-from nose_parameterized import parameterized
-from nose.tools import assert_greater, with_setup
+from nose.tools import with_setup
+
+import os
 
 
 @with_setup(setup, teardown)
 def test_catalog():
-
-    for klass, params in collect_meta_wrappers():
+    meta_wrappers = collect_meta_wrappers(
+        test_module=os.environ.get('TEST_MODULE', '').replace('.', os.path.sep),
+        test_all=os.environ.get('TEST_ALL', '') != ''
+    )
+    for klass, params in meta_wrappers:
         task = klass(**params)
         runtask(task, superclasses=[TagsTask, ColumnsTask, TableTask])
